refactor: move related changes filter to same directory
